MultiBonus.getAllItems() now working as intended
getAllItems() in MultiBonus now rightfully returns a List of HashMaps
with the possible Stats for the Item, that Class is for, so new Items
with the given Stats can now be created.
Also some QoL changes:
- Javadoc at Parser-Class
- Added some Lines at some Classes, to see which MultiBonus-Classes are
still missing. TODO: Add this to all Classes & Add all MultiBonuses
(This will take some time....)